Trim Reflect test harness
